GH-4097: Support "fuseki --version"
